[#401] Change how querying the baker key is handled
Problem: currently, we ask the user to input a tezos-client data
directory where the baker key is stored. This isn't the most
user-friendly way of doing things.

Solution: changed the Voting wizard to import the key into a directory
based on the network chosen. This key should persist between multiple
wizard invocation, so any subsequent voting should be easier.
